Moved .wrap, .before, and .after from Element to Node for flexibility. Overriding implementations in Element still return Element.
